+++ src/components/LayoutComposer/LayoutSideNavigation/LayoutWrapperAccountSettingsSideNav.js
@@ -138,6 +138,14 @@ const LayoutWrapperAccountSettingsSideNav = props => {
     },
     ...payoutDetailsMaybe,
     ...paymentMethodsMaybe,
+    {
+      text: <FormattedMessage id="LayoutWrapperAccountSettingsSideNav.manageAccountTabTitle" />,
+      selected: currentPage === 'ManageAccountPage',
+      id: 'ManageAccountPageTab',
+      linkProps: {
+        name: 'ManageAccountPage',
+      },
+    },
   ];
 
   return (