+++ src/containers/reducers.js
@@ -14,6 +14,7 @@ import PasswordChangePage from './PasswordChangePage/PasswordChangePage.duck';
 import PasswordRecoveryPage from './PasswordRecoveryPage/PasswordRecoveryPage.duck';
 import PasswordResetPage from './PasswordResetPage/PasswordResetPage.duck';
 import PaymentMethodsPage from './PaymentMethodsPage/PaymentMethodsPage.duck';
+import ManageAccountPage from './ManageAccountPage/ManageAccountPage.duck';
 import ProfilePage from './ProfilePage/ProfilePage.duck';
 import ProfileSettingsPage from './ProfileSettingsPage/ProfileSettingsPage.duck';
 import RequestQuotePage from './RequestQuotePage/RequestQuotePage.duck';
@@ -33,6 +34,7 @@ export {
   PasswordRecoveryPage,
   PasswordResetPage,
   PaymentMethodsPage,
+  ManageAccountPage,
   ProfilePage,
   ProfileSettingsPage,
   RequestQuotePage,